+++ src/main/java/org/filesys/server/filesys/db/DBDataInterface.java
@@ -36,12 +36,12 @@ public interface DBDataInterface {
     /**
      * Return the file data details for the specified file or stream.
      *
-     * @param fileId   int
-     * @param streamId int
+     * @param fileId   long
+     * @param streamId long
      * @return DBDataDetails
      * @exception DBException Database error
      */
-    public DBDataDetails getFileDataDetails(int fileId, int streamId)
+    public DBDataDetails getFileDataDetails(long fileId, long streamId)
             throws DBException;
 
     /**
@@ -54,37 +54,37 @@ public interface DBDataInterface {
     /**
      * Load file data from the database into a temporary/local file
      *
-     * @param fileId   int
-     * @param streamId int
+     * @param fileId   long
+     * @param streamId long
      * @param fileSeg  FileSegment
      * @exception DBException Database error
      * @exception IOException I/O error
      */
-    public void loadFileData(int fileId, int streamId, FileSegment fileSeg)
+    public void loadFileData(long fileId, long streamId, FileSegment fileSeg)
             throws DBException, IOException;
 
     /**
      * Load Jar file data from the database into a temporary file
      *
-     * @param jarId  int
+     * @param jarId  long
      * @param jarSeg FileSegment
      * @exception DBException Database error
      * @exception IOException I/O error
      */
-    public void loadJarData(int jarId, FileSegment jarSeg)
+    public void loadJarData(long jarId, FileSegment jarSeg)
             throws DBException, IOException;
 
     /**
      * Save the file data from the temporary/local file to the database
      *
-     * @param fileId   int
-     * @param streamId int
+     * @param fileId   long
+     * @param streamId long
      * @param fileSeg  FileSegment
      * @return int
      * @exception DBException Database error
      * @exception IOException I/O error
      */
-    public int saveFileData(int fileId, int streamId, FileSegment fileSeg)
+    public int saveFileData(long fileId, long streamId, FileSegment fileSeg)
             throws DBException, IOException;
 
     /**
@@ -102,21 +102,21 @@ public interface DBDataInterface {
     /**
      * Delete the file data for the specified file/stream
      *
-     * @param fileId   int
-     * @param streamId int
+     * @param fileId   long
+     * @param streamId long
      * @exception DBException Database error
      * @exception IOException I/O error
      */
-    public void deleteFileData(int fileId, int streamId)
+    public void deleteFileData(long fileId, long streamId)
             throws DBException, IOException;
 
     /**
      * Delete the file data for the specified Jar file
      *
-     * @param jarId int
+     * @param jarId long
      * @exception DBException Database error
      * @exception IOException I/O error
      */
-    public void deleteJarData(int jarId)
+    public void deleteJarData(long jarId)
             throws DBException, IOException;
 }
