+++ src/main/java/org/filesys/server/filesys/quota/QuotaManager.java
@@ -75,12 +75,12 @@ public interface QuotaManager {
      *
      * @param sess  SrvSession
      * @param tree  TreeConnection
-     * @param fid   int
+     * @param fid   long
      * @param path  String
      * @param alloc long
      * @throws IOException I/O error
      */
-    public void releaseSpace(SrvSession sess, TreeConnection tree, int fid, String path, long alloc)
+    public void releaseSpace(SrvSession sess, TreeConnection tree, long fid, String path, long alloc)
             throws IOException;
 
     /**
